Documents/8th sem/Data Privacy/Project/DPFedRec/PFedRec-Research/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
torch==2.2.1
numpy==1.26.4
pandas==2.2.1
requests==2.31.0
scikit-learn==1.4.1.post1
PyYAML==6.0.1
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.66.2
httpx==0.27.0
pytest==8.1.1
